requirements.txt
python-dotenv
aiohttp
jinja2
jsonschema
json-repair
tqdm
neo4j
pyyaml
yake
